refactor: split pipeline module into focused units
