+++ src-tauri/src/pipeline/test_utils.rs
@@ -1,0 +1,51 @@
+use crate::db::Task;
+
+pub fn make_task(retry_count: i64, pipeline_state: &str) -> Task {
+    Task {
+        id: "task-1".into(),
+        workspace_id: "ws-1".into(),
+        column_id: "col-1".into(),
+        title: "Test Task".into(),
+        description: None,
+        position: 0,
+        priority: "medium".into(),
+        agent_mode: None,
+        agent_status: None,
+        queued_at: None,
+        branch_name: None,
+        files_touched: "[]".into(),
+        checklist: None,
+        pipeline_state: pipeline_state.into(),
+        pipeline_triggered_at: None,
+        pipeline_error: None,
+        retry_count,
+        model: None,
+        agent_session_id: None,
+        last_script_exit_code: None,
+        review_status: None,
+        pr_number: None,
+        pr_url: None,
+        siege_iteration: 0,
+        siege_active: false,
+        siege_max_iterations: 5,
+        siege_last_checked: None,
+        pr_mergeable: None,
+        pr_ci_status: None,
+        pr_review_decision: None,
+        pr_comment_count: 0,
+        pr_is_draft: false,
+        pr_labels: "[]".into(),
+        pr_last_fetched: None,
+        pr_head_sha: None,
+        notify_stakeholders: None,
+        notification_sent_at: None,
+        trigger_overrides: None,
+        trigger_prompt: None,
+        last_output: None,
+        dependencies: None,
+        blocked: false,
+        worktree_path: None,
+        created_at: "2024-01-01T00:00:00Z".into(),
+        updated_at: "2024-01-01T00:00:00Z".into(),
+    }
+}
