+++ src-tauri/src/pipeline/triggers.rs
@@ -13,7 +13,10 @@ use std::collections::HashMap;
 use tauri::{AppHandle, Emitter};
 
 use super::template::{self, TemplateContext};
-use super::{emit_pipeline, PipelineState, EVT_TRIGGERED, EVT_RUNNING, EVT_ADVANCED};
+use super::{
+    emit_pipeline, emit_tasks_changed, fire_trigger, handle_trigger_failure, mark_complete,
+    PipelineEvent, PipelineState, EVT_ADVANCED, EVT_RUNNING, EVT_TRIGGERED,
+};
 
 // ─── V2 Trigger Types ─────────────────────────────────────────────────────
 
@@ -164,7 +167,13 @@ impl ResolvedStep {
 
     fn step_type(&self) -> &str {
         match self {
-            ResolvedStep::Shell { is_check, .. } => if *is_check { "check" } else { "bash" },
+            ResolvedStep::Shell { is_check, .. } => {
+                if *is_check {
+                    "check"
+                } else {
+                    "bash"
+                }
+            }
             ResolvedStep::Agent { .. } => "agent",
         }
     }
@@ -258,7 +267,14 @@ pub fn fire_on_entry(
         Option::None,
     )?;
 
-    emit_pipeline(app, EVT_TRIGGERED, &task.id, &column.id, PipelineState::Triggered, Some("V2 trigger fired".to_string()));
+    emit_pipeline(
+        app,
+        EVT_TRIGGERED,
+        &task.id,
+        &column.id,
+        PipelineState::Triggered,
+        Some("V2 trigger fired".to_string()),
+    );
 
     execute_action(conn, app, &updated_task, column, &action, prev_column)
 }
@@ -297,15 +313,43 @@ fn execute_action(
     other_column: Option<&Column>,
 ) -> Result<Task, AppError> {
     match action {
-        TriggerActionV2::SpawnCli { cli, command, prompt_template, prompt, flags, use_queue: _, model } => {
-            execute_spawn_cli(conn, app, task, column, other_column, cli.as_deref(), command.as_deref(), prompt_template.as_deref(), prompt.as_deref(), flags.as_deref(), model.as_deref())
-        }
-        TriggerActionV2::MoveColumn { target } => {
-            execute_move_column(conn, app, task, target)
-        }
-        TriggerActionV2::TriggerTask { target_task, action: task_action, target_column, inject_prompt } => {
-            execute_trigger_task(conn, app, task, column, target_task, task_action, target_column.as_deref(), inject_prompt.as_deref())
-        }
+        TriggerActionV2::SpawnCli {
+            cli,
+            command,
+            prompt_template,
+            prompt,
+            flags,
+            use_queue: _,
+            model,
+        } => execute_spawn_cli(
+            conn,
+            app,
+            task,
+            column,
+            other_column,
+            cli.as_deref(),
+            command.as_deref(),
+            prompt_template.as_deref(),
+            prompt.as_deref(),
+            flags.as_deref(),
+            model.as_deref(),
+        ),
+        TriggerActionV2::MoveColumn { target } => execute_move_column(conn, app, task, target),
+        TriggerActionV2::TriggerTask {
+            target_task,
+            action: task_action,
+            target_column,
+            inject_prompt,
+        } => execute_trigger_task(
+            conn,
+            app,
+            task,
+            column,
+            target_task,
+            task_action,
+            target_column.as_deref(),
+            inject_prompt.as_deref(),
+        ),
         TriggerActionV2::RunScript { script_id } => {
             execute_run_script(conn, app, task, column, other_column, script_id)
         }
@@ -342,12 +386,20 @@ fn ensure_task_worktree(
         match branch_manager::create_task_branch(repo_path, &slug, None) {
             Ok(branch_name) => {
                 task = db::update_task_branch(conn, &task.id, Some(&branch_name))?;
-                log::info!("[triggers] Auto-created branch '{}' for task {}", branch_name, task.id);
+                log::info!(
+                    "[triggers] Auto-created branch '{}' for task {}",
+                    branch_name,
+                    task.id
+                );
             }
             Err(e) => {
                 // Branch may already exist (e.g. from a previous attempt)
                 let branch_name = format!("bentoya/{}", slug);
-                log::warn!("[triggers] Branch creation failed ({}), trying existing '{}'", e, branch_name);
+                log::warn!(
+                    "[triggers] Branch creation failed ({}), trying existing '{}'",
+                    e,
+                    branch_name
+                );
                 task = db::update_task_branch(conn, &task.id, Some(&branch_name))?;
             }
         }
@@ -355,7 +407,10 @@ fn ensure_task_worktree(
 
     let branch_name = task.branch_name.as_deref().unwrap_or("");
     if branch_name.is_empty() {
-        log::warn!("[triggers] Could not determine branch for task {}, skipping worktree", task.id);
+        log::warn!(
+            "[triggers] Could not determine branch for task {}, skipping worktree",
+            task.id
+        );
         return Ok(task);
     }
 
@@ -363,11 +418,19 @@ fn ensure_task_worktree(
     match branch_manager::create_task_worktree(repo_path, branch_name, &task.id) {
         Ok(wt_path) => {
             task = db::update_task_worktree_path(conn, &task.id, Some(&wt_path))?;
-            super::emit_tasks_changed(app, &task.workspace_id, "worktree_auto_created");
-            log::info!("[triggers] Auto-created worktree at '{}' for task {}", wt_path, task.id);
+            emit_tasks_changed(app, &task.workspace_id, "worktree_auto_created");
+            log::info!(
+                "[triggers] Auto-created worktree at '{}' for task {}",
+                wt_path,
+                task.id
+            );
         }
         Err(e) => {
-            log::error!("[triggers] Failed to create worktree for task {}: {}", task.id, e);
+            log::error!(
+                "[triggers] Failed to create worktree for task {}: {}",
+                task.id,
+                e
+            );
             // Continue without worktree — agent falls back to repo root
         }
     }
@@ -402,13 +465,18 @@ fn execute_spawn_cli(
     let working_dir = resolve_working_dir(&task, &workspace.repo_path);
 
     if !working_dir.is_empty() && !std::path::Path::new(&working_dir).exists() {
-        log::warn!("Working dir '{}' does not exist, agent may fail", working_dir);
+        log::warn!(
+            "Working dir '{}' does not exist, agent may fail",
+            working_dir
+        );
     }
 
     // Write .task.md to worktree — agent reads this instead of getting full spec in prompt
     if !working_dir.is_empty() {
         let task_md_path = std::path::Path::new(&working_dir).join(".task.md");
-        let checklist_section = task.checklist.as_deref()
+        let checklist_section = task
+            .checklist
+            .as_deref()
             .filter(|c| !c.is_empty() && *c != "[]")
             .map(|c| format!("\n## Checklist\n{}\n", c))
             .unwrap_or_default();
@@ -428,49 +496,52 @@ fn execute_spawn_cli(
         }
 
         // Exclude .task.md from git (avoid agent committing it)
-        let exclude_path = std::path::Path::new(&working_dir).join(".git").join("info").join("exclude");
+        let exclude_path = std::path::Path::new(&working_dir)
+            .join(".git")
+            .join("info")
+            .join("exclude");
         if exclude_path.exists() {
             if let Ok(content) = std::fs::read_to_string(&exclude_path) {
                 if !content.contains(".task.md") {
-                    let _ = std::fs::write(&exclude_path, format!("{}\n.task.md\n.task-handoff.md\n", content.trim_end()));
+                    let _ = std::fs::write(
+                        &exclude_path,
+                        format!("{}\n.task.md\n.task-handoff.md\n", content.trim_end()),
+                    );
                 }
             }
         }
     }
 
     let ctx = TemplateContext {
-        task: &task, column, workspace: &workspace,
-        prev_column: other_column, next_column: Option::None, dep_tasks: HashMap::new(),
+        task: &task,
+        column,
+        workspace: &workspace,
+        prev_column: other_column,
+        next_column: Option::None,
+        dep_tasks: HashMap::new(),
     };
 
     // Resolve prompt: direct prompt > template > .task.md pointer (token-optimized default)
     let resolved_prompt = if let Some(p) = prompt {
-        if !p.is_empty() { template::interpolate(p, &ctx) }
-        else if let Some(tmpl) = prompt_template { template::interpolate(tmpl, &ctx) }
-        else { format!("{}\n\nSee .task.md for full spec.", task.title) }
+        if !p.is_empty() {
+            template::interpolate(p, &ctx)
+        } else if let Some(tmpl) = prompt_template {
+            template::interpolate(tmpl, &ctx)
+        } else {
+            format!("{}\n\nSee .task.md for full spec.", task.title)
+        }
     } else if let Some(tmpl) = prompt_template {
         template::interpolate(tmpl, &ctx)
     } else {
         format!("{}\n\nSee .task.md for full spec.", task.title)
     };
 
-    // Resolve CLI and model from workspace config (both use the same parsed value)
-    let workspace_config: serde_json::Value = serde_json::from_str(&workspace.config).unwrap_or_default();
-
-    // Resolve CLI: trigger config > workspace default > "claude"
-    let ws_default_cli = workspace_config.get("defaultAgentCli").and_then(|v| v.as_str()).filter(|s| !s.is_empty());
-    let cli_type = cli.or(ws_default_cli).unwrap_or("claude").to_string();
-
-    // Resolve model: task override > trigger config > workspace default > none
-    let ws_default_model = workspace_config.get("defaultModel").and_then(|v| v.as_str()).filter(|s| !s.is_empty());
-    let resolved_model = task.model.as_deref().or(model).or(ws_default_model).map(|m| m.to_string());
     // Resolve CLI: trigger config > workspace config > global settings
     let cli_type = cli.map(|c| c.to_string()).unwrap_or_else(|| {
         let settings = crate::config::AppSettings::load();
-        settings.resolve_with_workspace(
-            Some(&workspace.config),
-            "default_agent_cli",
-        ).unwrap_or_else(|| settings.default_agent_cli.clone())
+        settings
+            .resolve_with_workspace(Some(&workspace.config), "default_agent_cli")
+            .unwrap_or_else(|| settings.default_agent_cli.clone())
     });
 
     // Prepend slash command if provided
@@ -490,7 +561,11 @@ fn execute_spawn_cli(
     }
 
     let updated_task = db::update_task_pipeline_state(
-        conn, &task.id, PipelineState::Running.as_str(), Some(&ts), Option::None,
+        conn,
+        &task.id,
+        PipelineState::Running.as_str(),
+        Some(&ts),
+        Option::None,
     )?;
 
     // Build env vars
@@ -504,15 +579,32 @@ fn execute_spawn_cli(
         env_vars.insert("TRIGGER_FLAGS".to_string(), f.join(" "));
     }
 
-    emit_pipeline(app, EVT_RUNNING, &task.id, &column.id, PipelineState::Running, Some(format!("CLI trigger: {}", cli_type)));
+    emit_pipeline(
+        app,
+        EVT_RUNNING,
+        &task.id,
+        &column.id,
+        PipelineState::Running,
+        Some(format!("CLI trigger: {}", cli_type)),
+    );
 
     // Resolve model: task override > trigger config > workspace config > global settings
-    let resolved_model = task.model.as_deref().or(model).map(|m| m.to_string()).or_else(|| {
-        let settings = crate::config::AppSettings::load();
-        let m = settings.resolve_with_workspace(Some(&workspace.config), "default_model")
-            .unwrap_or_else(|| settings.default_model.clone());
-        if m.is_empty() { None } else { Some(m) }
-    });
+    let resolved_model = task
+        .model
+        .as_deref()
+        .or(model)
+        .map(|m| m.to_string())
+        .or_else(|| {
+            let settings = crate::config::AppSettings::load();
+            let m = settings
+                .resolve_with_workspace(Some(&workspace.config), "default_model")
+                .unwrap_or_else(|| settings.default_model.clone());
+            if m.is_empty() {
+                None
+            } else {
+                Some(m)
+            }
+        });
     let mut cli_args = Vec::new();
     if let Some(ref m) = resolved_model {
         cli_args.push("--model".to_string());
@@ -520,8 +612,13 @@ fn execute_spawn_cli(
     }
 
     bridge::spawn_cli_trigger_task(
-        app.clone(), task.id.clone(), cli_type, cli_args,
-        working_dir, initial_prompt, Some(env_vars),
+        app.clone(),
+        task.id.clone(),
+        cli_type,
+        cli_args,
+        working_dir,
+        initial_prompt,
+        Some(env_vars),
     );
 
     Ok(updated_task)
@@ -539,7 +636,8 @@ fn execute_move_column(
         let max_pos: i64 = conn
             .query_row(
                 "SELECT COALESCE(MAX(position), -1) FROM tasks WHERE column_id = ?1",
-                rusqlite::params![col.id], |row| row.get(0),
+                rusqlite::params![col.id],
+                |row| row.get(0),
             )
             .unwrap_or(-1);
 
@@ -548,11 +646,18 @@ fn execute_move_column(
             rusqlite::params![col.id, max_pos + 1, ts, task.id],
         ).map_err(AppError::from)?;
 
-        emit_pipeline(app, EVT_ADVANCED, &task.id, &col.id, PipelineState::Idle, Some(format!("Moved to {}", col.name)));
+        emit_pipeline(
+            app,
+            EVT_ADVANCED,
+            &task.id,
+            &col.id,
+            PipelineState::Idle,
+            Some(format!("Moved to {}", col.name)),
+        );
 
         let moved_task = db::get_task(conn, &task.id)?;
-        super::emit_tasks_changed(app, &task.workspace_id, "trigger_move_column");
-        let _ = super::fire_trigger(conn, app, &moved_task, &col);
+        emit_tasks_changed(app, &task.workspace_id, "trigger_move_column");
+        let _ = fire_trigger(conn, app, &moved_task, &col);
         return Ok(db::get_task(conn, &task.id)?);
     }
     Ok(task.clone())
@@ -591,11 +696,12 @@ fn execute_trigger_task(
                 conn.execute(
                     "UPDATE tasks SET blocked = 0, updated_at = ?1 WHERE id = ?2",
                     rusqlite::params![db::now(), target.id],
-                ).map_err(AppError::from)?;
+                )
+                .map_err(AppError::from)?;
             }
             TriggerTaskActionTypeV2::Start => {
                 let col = db::get_column(conn, &target.column_id)?;
-                let _ = super::fire_trigger(conn, app, &target, &col);
+                let _ = fire_trigger(conn, app, &target, &col);
             }
         }
 
@@ -603,14 +709,19 @@ fn execute_trigger_task(
         if let Some(inject) = inject_prompt {
             let workspace = db::get_workspace(conn, &task.workspace_id)?;
             let ctx = TemplateContext {
-                task, column, workspace: &workspace,
-                prev_column: Option::None, next_column: Option::None, dep_tasks: HashMap::new(),
+                task,
+                column,
+                workspace: &workspace,
+                prev_column: Option::None,
+                next_column: Option::None,
+                dep_tasks: HashMap::new(),
             };
             let resolved = template::interpolate(inject, &ctx);
             conn.execute(
                 "UPDATE tasks SET trigger_prompt = ?1, updated_at = ?2 WHERE id = ?3",
                 rusqlite::params![resolved, db::now(), target.id],
-            ).map_err(AppError::from)?;
+            )
+            .map_err(AppError::from)?;
         }
     }
     Ok(task.clone())
@@ -628,17 +739,28 @@ fn execute_create_pr(
     let branch_name = match &task.branch_name {
         Some(b) if !b.is_empty() => b.clone(),
         _ => {
-            return super::handle_trigger_failure(
-                conn, app, task, column,
+            return handle_trigger_failure(
+                conn,
+                app,
+                task,
+                column,
                 "Cannot create PR: task has no branch_name",
             );
         }
     };
 
-    if task.pr_number.is_some() {
-        log::info!("[create_pr] Task {} already has PR #{}, skipping", task.id, task.pr_number.unwrap());
+    if let Some(pr_number) = task.pr_number {
+        log::info!(
+            "[create_pr] Task {} already has PR #{}, skipping",
+            task.id,
+            pr_number
+        );
         let updated = db::update_task_pipeline_state(
-            conn, &task.id, PipelineState::Idle.as_str(), None, None,
+            conn,
+            &task.id,
+            PipelineState::Idle.as_str(),
+            None,
+            None,
         )?;
         return Ok(updated);
     }
@@ -651,11 +773,21 @@ fn execute_create_pr(
     let column_id = column.id.clone();
 
     let updated_task = db::update_task_pipeline_state(
-        conn, &task.id, PipelineState::Running.as_str(),
-        Some(&db::now()), None,
+        conn,
+        &task.id,
+        PipelineState::Running.as_str(),
+        Some(&db::now()),
+        None,
     )?;
 
-    emit_pipeline(app, EVT_RUNNING, &task.id, &column.id, PipelineState::Running, Some("Creating PR".to_string()));
+    emit_pipeline(
+        app,
+        EVT_RUNNING,
+        &task.id,
+        &column.id,
+        PipelineState::Running,
+        Some("Creating PR".to_string()),
+    );
 
     let app_handle = app.clone();
 
@@ -675,7 +807,10 @@ fn execute_create_pr(
                     // Fallback: force-push if regular push fails (e.g. stale remote branch)
                     // Safety: only force-push bentoya/* branches, never main/master
                     if branch_name.starts_with("bentoya/") {
-                        log::warn!("[pipeline] Regular push failed, trying force-push: {}", stderr.trim());
+                        log::warn!(
+                            "[pipeline] Regular push failed, trying force-push: {}",
+                            stderr.trim()
+                        );
                         let force_output = std::process::Command::new("git")
                             .args(["push", "-u", "origin", &branch_name, "--force"])
                             .current_dir(&repo_path)
@@ -683,21 +818,33 @@ fn execute_create_pr(
                             .map_err(|e| format!("Failed to force-push branch: {}", e))?;
                         if !force_output.status.success() {
                             let force_stderr = String::from_utf8_lossy(&force_output.stderr);
-                            return Err(format!("git push --force failed: {}", force_stderr.trim()));
+                            return Err(format!(
+                                "git push --force failed: {}",
+                                force_stderr.trim()
+                            ));
                         }
                     } else {
-                        return Err(format!("git push failed (force-push not allowed on '{}'): {}", branch_name, stderr.trim()));
+                        return Err(format!(
+                            "git push failed (force-push not allowed on '{}'): {}",
+                            branch_name,
+                            stderr.trim()
+                        ));
                     }
                 }
             }
 
             let output = std::process::Command::new("gh")
                 .args([
-                    "pr", "create",
-                    "--title", &pr_title,
-                    "--body", &pr_body,
-                    "--base", &base,
-                    "--head", &branch_name,
+                    "pr",
+                    "create",
+                    "--title",
+                    &pr_title,
+                    "--body",
+                    &pr_body,
+                    "--base",
+                    &base,
+                    "--head",
+                    &branch_name,
                 ])
                 .current_dir(&repo_path)
                 .output()
@@ -717,17 +864,29 @@ fn execute_create_pr(
                 .ok_or_else(|| format!("Failed to parse PR number from URL: {}", pr_url))?;
 
             Ok((pr_number, pr_url))
-        }).await;
+        })
+        .await;
 
         let db_path = crate::db::db_path();
         let conn = match rusqlite::Connection::open(&db_path) {
-            Ok(c) => { let _ = c.execute_batch("PRAGMA journal_mode=WAL;"); Some(c) }
-            Err(e) => { log::error!("[create_pr] Failed to open DB: {}", e); None }
+            Ok(c) => {
+                let _ = c.execute_batch("PRAGMA journal_mode=WAL;");
+                Some(c)
+            }
+            Err(e) => {
+                log::error!("[create_pr] Failed to open DB: {}", e);
+                None
+            }
         };
 
         let success = match result {
             Ok(Ok((pr_number, pr_url))) => {
-                log::info!("[create_pr] Created PR #{} for task {}: {}", pr_number, task_id, pr_url);
+                log::info!(
+                    "[create_pr] Created PR #{} for task {}: {}",
+                    pr_number,
+                    task_id,
+                    pr_url
+                );
                 if let Some(ref conn) = conn {
                     let _ = db::update_task_pr_info(conn, &task_id, Some(pr_number), Some(&pr_url));
                 }
@@ -735,13 +894,16 @@ fn execute_create_pr(
             }
             Ok(Err(e)) => {
                 log::error!("[create_pr] Failed for task {}: {}", task_id, e);
-                let _ = app_handle.emit("pipeline:error", &super::PipelineEvent {
-                    task_id: task_id.clone(),
-                    column_id: column_id.clone(),
-                    event_type: "error".to_string(),
-                    state: PipelineState::Idle.as_str().to_string(),
-                    message: Some(e),
-                });
+                let _ = app_handle.emit(
+                    "pipeline:error",
+                    &PipelineEvent {
+                        task_id: task_id.clone(),
+                        column_id: column_id.clone(),
+                        event_type: "error".to_string(),
+                        state: PipelineState::Idle.as_str().to_string(),
+                        message: Some(e),
+                    },
+                );
                 false
             }
             Err(e) => {
@@ -752,7 +914,7 @@ fn execute_create_pr(
 
         // Mark complete so pipeline can advance (also emits tasks:changed)
         if let Some(conn) = conn {
-            if let Err(e) = super::mark_complete(&conn, &app_handle, &task_id, success) {
+            if let Err(e) = mark_complete(&conn, &app_handle, &task_id, success) {
                 log::error!("[create_pr] mark_complete failed: {}", e);
             }
         }
@@ -769,213 +931,266 @@ fn execute_run_script(
     other_column: Option<&Column>,
     script_id: &str,
 ) -> Result<Task, AppError> {
-            // Load script from DB and execute steps sequentially
-            let script = db::get_script(conn, script_id)
-                .map_err(|_| AppError::NotFound(format!("Script '{}' not found", script_id)))?;
-            let workspace = db::get_workspace(conn, &task.workspace_id)?;
-            let working_dir = resolve_working_dir(task, &workspace.repo_path);
+    // Load script from DB and execute steps sequentially
+    let script = db::get_script(conn, script_id)
+        .map_err(|_| AppError::NotFound(format!("Script '{}' not found", script_id)))?;
+    let workspace = db::get_workspace(conn, &task.workspace_id)?;
+    let working_dir = resolve_working_dir(task, &workspace.repo_path);
 
-            // Validate working dir exists
-            if !working_dir.is_empty() && !std::path::Path::new(&working_dir).exists() {
-                log::warn!("Working dir '{}' does not exist, script may fail", working_dir);
-            }
+    // Validate working dir exists
+    if !working_dir.is_empty() && !std::path::Path::new(&working_dir).exists() {
+        log::warn!(
+            "Working dir '{}' does not exist, script may fail",
+            working_dir
+        );
+    }
 
-            let ts = db::now();
-            let updated_task = db::update_task_pipeline_state(
-                conn,
-                &task.id,
-                PipelineState::Running.as_str(),
-                Some(&ts),
-                Option::None,
-            )?;
+    let ts = db::now();
+    let updated_task = db::update_task_pipeline_state(
+        conn,
+        &task.id,
+        PipelineState::Running.as_str(),
+        Some(&ts),
+        Option::None,
+    )?;
 
-            emit_pipeline(app, EVT_RUNNING, &task.id, &column.id, PipelineState::Running, Some(format!("Running script: {}", script.name)));
+    emit_pipeline(
+        app,
+        EVT_RUNNING,
+        &task.id,
+        &column.id,
+        PipelineState::Running,
+        Some(format!("Running script: {}", script.name)),
+    );
 
-            // Parse steps
-            let steps: Vec<serde_json::Value> = serde_json::from_str(&script.steps)
-                .unwrap_or_default();
+    // Parse steps
+    let steps: Vec<serde_json::Value> = serde_json::from_str(&script.steps).unwrap_or_default();
 
-            // Pre-interpolate all template variables before moving into async block
-            let ctx = TemplateContext {
-                task,
-                column,
-                workspace: &workspace,
-                prev_column: other_column,
-                next_column: Option::None,
-                dep_tasks: HashMap::new(),
-            };
+    // Pre-interpolate all template variables before moving into async block
+    let ctx = TemplateContext {
+        task,
+        column,
+        workspace: &workspace,
+        prev_column: other_column,
+        next_column: Option::None,
+        dep_tasks: HashMap::new(),
+    };
 
-            // Resolve all steps into owned data
-            let resolved_steps: Vec<ResolvedStep> = steps.iter().map(|step| {
-                let step_type = step.get("type").and_then(|v| v.as_str()).unwrap_or("bash").to_string();
-                let step_name = step.get("name").and_then(|v| v.as_str()).unwrap_or("Step").to_string();
+    // Resolve all steps into owned data
+    let resolved_steps: Vec<ResolvedStep> = steps
+        .iter()
+        .map(|step| {
+            let step_type = step
+                .get("type")
+                .and_then(|v| v.as_str())
+                .unwrap_or("bash")
+                .to_string();
+            let step_name = step
+                .get("name")
+                .and_then(|v| v.as_str())
+                .unwrap_or("Step")
+                .to_string();
 
-                match step_type.as_str() {
-                    "bash" | "check" => {
-                        let command = step.get("command").and_then(|v| v.as_str()).unwrap_or("");
-                        let command = template::interpolate(command, &ctx);
-                        let work_dir = step.get("workDir")
-                            .and_then(|v| v.as_str())
-                            .map(|d| template::interpolate(d, &ctx))
-                            .unwrap_or_else(|| working_dir.clone());
-                        let continue_on_error = step.get("continueOnError")
-                            .and_then(|v| v.as_bool())
-                            .unwrap_or(false);
-                        let fail_message = step.get("failMessage")
-                            .and_then(|v| v.as_str())
-                            .map(|s| s.to_string());
+            match step_type.as_str() {
+                "bash" | "check" => {
+                    let command = step.get("command").and_then(|v| v.as_str()).unwrap_or("");
+                    let command = template::interpolate(command, &ctx);
+                    let work_dir = step
+                        .get("workDir")
+                        .and_then(|v| v.as_str())
+                        .map(|d| template::interpolate(d, &ctx))
+                        .unwrap_or_else(|| working_dir.clone());
+                    let continue_on_error = step
+                        .get("continueOnError")
+                        .and_then(|v| v.as_bool())
+                        .unwrap_or(false);
+                    let fail_message = step
+                        .get("failMessage")
+                        .and_then(|v| v.as_str())
+                        .map(|s| s.to_string());
 
-                        ResolvedStep::Shell {
-                            name: step_name,
-                            is_check: step_type == "check",
-                            command,
-                            work_dir,
-                            continue_on_error,
-                            fail_message,
-                        }
-                    }
-                    "agent" => {
-                        let prompt = step.get("prompt").and_then(|v| v.as_str()).unwrap_or("");
-                        let prompt = template::interpolate(prompt, &ctx);
-                        let model = step.get("model").and_then(|v| v.as_str()).map(|s| s.to_string());
-                        let command = step.get("command").and_then(|v| v.as_str()).map(|s| s.to_string());
-
-                        ResolvedStep::Agent {
-                            name: step_name,
-                            prompt,
-                            model,
-                            command,
-                        }
-                    }
-                    _ => ResolvedStep::Shell {
+                    ResolvedStep::Shell {
                         name: step_name,
-                        is_check: false,
-                        command: String::new(),
-                        work_dir: working_dir.clone(),
-                        continue_on_error: true,
-                        fail_message: None,
-                    },
+                        is_check: step_type == "check",
+                        command,
+                        work_dir,
+                        continue_on_error,
+                        fail_message,
+                    }
                 }
-            }).collect();
+                "agent" => {
+                    let prompt = step.get("prompt").and_then(|v| v.as_str()).unwrap_or("");
+                    let prompt = template::interpolate(prompt, &ctx);
+                    let model = step
+                        .get("model")
+                        .and_then(|v| v.as_str())
+                        .map(|s| s.to_string());
+                    let command = step
+                        .get("command")
+                        .and_then(|v| v.as_str())
+                        .map(|s| s.to_string());
 
-            let task_id = task.id.clone();
-            let app_handle = app.clone();
-            let column_id = column.id.clone();
-            let workspace_path = working_dir.clone();
+                    ResolvedStep::Agent {
+                        name: step_name,
+                        prompt,
+                        model,
+                        command,
+                    }
+                }
+                _ => ResolvedStep::Shell {
+                    name: step_name,
+                    is_check: false,
+                    command: String::new(),
+                    work_dir: working_dir.clone(),
+                    continue_on_error: true,
+                    fail_message: None,
+                },
+            }
+        })
+        .collect();
 
-            // Execute steps in a background task (all data is owned)
-            tokio::spawn(async move {
-                let mut success = true;
-                let total = resolved_steps.len();
+    let task_id = task.id.clone();
+    let app_handle = app.clone();
+    let column_id = column.id.clone();
+    let workspace_path = working_dir.clone();
 
-                for (i, step) in resolved_steps.iter().enumerate() {
-                    let step_name = step.name();
+    // Execute steps in a background task (all data is owned)
+    tokio::spawn(async move {
+        let mut success = true;
+        let total = resolved_steps.len();
 
-                    // Emit progress
-                    let _ = app_handle.emit(
-                        "pipeline:step_progress",
-                        &serde_json::json!({
-                            "taskId": task_id,
-                            "columnId": column_id,
-                            "step": i + 1,
-                            "total": total,
-                            "name": step_name,
-                            "type": step.step_type(),
-                        }),
+        for (i, step) in resolved_steps.iter().enumerate() {
+            let step_name = step.name();
+
+            // Emit progress
+            let _ = app_handle.emit(
+                "pipeline:step_progress",
+                &serde_json::json!({
+                    "taskId": task_id,
+                    "columnId": column_id,
+                    "step": i + 1,
+                    "total": total,
+                    "name": step_name,
+                    "type": step.step_type(),
+                }),
+            );
+
+            match step {
+                ResolvedStep::Shell {
+                    name,
+                    is_check,
+                    command,
+                    work_dir,
+                    continue_on_error,
+                    fail_message,
+                } => {
+                    let output = tokio::process::Command::new("sh")
+                        .arg("-c")
+                        .arg(command)
+                        .current_dir(work_dir)
+                        .output()
+                        .await;
+
+                    match output {
+                        Ok(out) => {
+                            let stdout = String::from_utf8_lossy(&out.stdout);
+                            let stderr = String::from_utf8_lossy(&out.stderr);
+                            log::info!(
+                                "[script:{}] Step {}/{} '{}': exit={}, stdout={} bytes",
+                                task_id,
+                                i + 1,
+                                total,
+                                name,
+                                out.status.code().unwrap_or(-1),
+                                stdout.len()
+                            );
+                            if !stderr.is_empty() {
+                                log::warn!(
+                                    "[script:{}] stderr: {}",
+                                    task_id,
+                                    stderr.chars().take(500).collect::<String>()
+                                );
+                            }
+                            if !out.status.success() {
+                                if *is_check {
+                                    let msg = fail_message.as_deref().unwrap_or("Check failed");
+                                    log::warn!("[script:{}] Check failed: {}", task_id, msg);
+                                }
+                                if !continue_on_error {
+                                    success = false;
+                                    break;
+                                }
+                            }
+                        }
+                        Err(e) => {
+                            log::error!(
+                                "[script:{}] Failed to run step '{}': {}",
+                                task_id,
+                                name,
+                                e
+                            );
+                            if !continue_on_error {
+                                success = false;
+                                break;
+                            }
+                        }
+                    }
+                }
+                ResolvedStep::Agent {
+                    prompt,
+                    model,
+                    command,
+                    ..
+                } => {
+                    let initial_prompt = if let Some(cmd) = command {
+                        if prompt.is_empty() {
+                            cmd.clone()
+                        } else {
+                            format!("{}\n\n{}", cmd, prompt)
+                        }
+                    } else {
+                        prompt.clone()
+                    };
+
+                    let mut cli_args = Vec::new();
+                    if let Some(m) = model {
+                        cli_args.push("--model".to_string());
+                        cli_args.push(m.clone());
+                    }
+
+                    // Spawn CLI — mark_complete called by PTY exit handler
+                    bridge::spawn_cli_trigger_task(
+                        app_handle.clone(),
+                        task_id.clone(),
+                        "claude".to_string(),
+                        cli_args,
+                        workspace_path.clone(),
+                        initial_prompt,
+                        Option::None,
                     );
 
-                    match step {
-                        ResolvedStep::Shell { name, is_check, command, work_dir, continue_on_error, fail_message } => {
-                            let output = tokio::process::Command::new("sh")
-                                .arg("-c")
-                                .arg(command)
-                                .current_dir(work_dir)
-                                .output()
-                                .await;
-
-                            match output {
-                                Ok(out) => {
-                                    let stdout = String::from_utf8_lossy(&out.stdout);
-                                    let stderr = String::from_utf8_lossy(&out.stderr);
-                                    log::info!(
-                                        "[script:{}] Step {}/{} '{}': exit={}, stdout={} bytes",
-                                        task_id, i + 1, total, name,
-                                        out.status.code().unwrap_or(-1),
-                                        stdout.len()
-                                    );
-                                    if !stderr.is_empty() {
-                                        log::warn!("[script:{}] stderr: {}", task_id, stderr.chars().take(500).collect::<String>());
-                                    }
-                                    if !out.status.success() {
-                                        if *is_check {
-                                            let msg = fail_message.as_deref().unwrap_or("Check failed");
-                                            log::warn!("[script:{}] Check failed: {}", task_id, msg);
-                                        }
-                                        if !continue_on_error {
-                                            success = false;
-                                            break;
-                                        }
-                                    }
-                                }
-                                Err(e) => {
-                                    log::error!("[script:{}] Failed to run step '{}': {}", task_id, name, e);
-                                    if !continue_on_error {
-                                        success = false;
-                                        break;
-                                    }
-                                }
-                            }
-                        }
-                        ResolvedStep::Agent { prompt, model, command, .. } => {
-                            let initial_prompt = if let Some(cmd) = command {
-                                if prompt.is_empty() {
-                                    cmd.clone()
-                                } else {
-                                    format!("{}\n\n{}", cmd, prompt)
-                                }
-                            } else {
-                                prompt.clone()
-                            };
-
-                            let mut cli_args = Vec::new();
-                            if let Some(m) = model {
-                                cli_args.push("--model".to_string());
-                                cli_args.push(m.clone());
-                            }
-
-                            // Spawn CLI — mark_complete called by PTY exit handler
-                            bridge::spawn_cli_trigger_task(
-                                app_handle.clone(),
-                                task_id.clone(),
-                                "claude".to_string(),
-                                cli_args,
-                                workspace_path.clone(),
-                                initial_prompt,
-                                Option::None,
-                            );
-
-                            // Agent step hands off to PTY exit handler
-                            return;
-                        }
-                    }
+                    // Agent step hands off to PTY exit handler
+                    return;
                 }
+            }
+        }
 
-                // All bash/check steps done — call mark_complete
-                let db_path = crate::db::db_path();
-                match rusqlite::Connection::open(&db_path) {
-                    Ok(conn) => {
-                        let _ = conn.execute_batch("PRAGMA journal_mode=WAL;");
-                        if let Err(e) = super::mark_complete(&conn, &app_handle, &task_id, success) {
-                            log::error!("[script:{}] mark_complete failed: {}", task_id, e);
-                        }
-                    }
-                    Err(e) => {
-                        log::error!("[script:{}] Failed to open DB for mark_complete: {} — task stuck in running state", task_id, e);
-                    }
+        // All bash/check steps done — call mark_complete
+        let db_path = crate::db::db_path();
+        match rusqlite::Connection::open(&db_path) {
+            Ok(conn) => {
+                let _ = conn.execute_batch("PRAGMA journal_mode=WAL;");
+                if let Err(e) = mark_complete(&conn, &app_handle, &task_id, success) {
+                    log::error!("[script:{}] mark_complete failed: {}", task_id, e);
                 }
-            });
+            }
+            Err(e) => {
+                log::error!("[script:{}] Failed to open DB for mark_complete: {} — task stuck in running state", task_id, e);
+            }
+        }
+    });
 
-            Ok(updated_task)
+    Ok(updated_task)
 }
 
 // Note: dependency task interpolation ({dep.<id>.title}) is handled at the
@@ -990,7 +1205,11 @@ pub fn resolve_column_target(
     let current_col = db::get_column(conn, &task.column_id)?;
 
     match target {
-        "next" => Ok(db::get_next_column(conn, &task.workspace_id, current_col.position)?),
+        "next" => Ok(db::get_next_column(
+            conn,
+            &task.workspace_id,
+            current_col.position,
+        )?),
         "previous" => {
             if current_col.position > 0 {
                 let cols = db::list_columns(conn, &task.workspace_id)?;
@@ -1012,7 +1231,13 @@ mod tests {
     use super::*;
     use crate::db;
 
-    fn setup() -> (rusqlite::Connection, db::Workspace, db::Column, db::Column, db::Column) {
+    fn setup() -> (
+        rusqlite::Connection,
+        db::Workspace,
+        db::Column,
+        db::Column,
+        db::Column,
+    ) {
         let conn = db::init_test().unwrap();
         let ws = db::insert_workspace(&conn, "Test", "/tmp/test").unwrap();
         let col1 = db::insert_column(&conn, &ws.id, "Backlog", 0).unwrap();
@@ -1086,7 +1311,8 @@ mod tests {
         conn.execute(
             "UPDATE tasks SET trigger_overrides = ?1 WHERE id = ?2",
             rusqlite::params![r#"{"skip_triggers": true}"#, task.id],
-        ).unwrap();
+        )
+        .unwrap();
         let task = db::get_task(&conn, &task.id).unwrap();
 
         let triggers = ColumnTriggersV2 {
@@ -1104,7 +1330,10 @@ mod tests {
         };
 
         let result = resolve_trigger(&triggers, &task, "on_entry");
-        assert!(result.is_none(), "skip_triggers should suppress the trigger");
+        assert!(
+            result.is_none(),
+            "skip_triggers should suppress the trigger"
+        );
     }
 
     #[test]
@@ -1116,7 +1345,8 @@ mod tests {
         conn.execute(
             "UPDATE tasks SET trigger_overrides = ?1 WHERE id = ?2",
             rusqlite::params![r#"{"on_entry": {"model": "opus"}}"#, task.id],
-        ).unwrap();
+        )
+        .unwrap();
         let task = db::get_task(&conn, &task.id).unwrap();
 
         let triggers = ColumnTriggersV2 {
@@ -1137,7 +1367,11 @@ mod tests {
         assert!(result.is_some());
         if let Some(TriggerActionV2::SpawnCli { model, command, .. }) = result {
             assert_eq!(model.as_deref(), Some("opus"), "task override should win");
-            assert_eq!(command.as_deref(), Some("/start-task"), "base command preserved");
+            assert_eq!(
+                command.as_deref(),
+                Some("/start-task"),
+                "base command preserved"
+            );
         } else {
             panic!("Expected SpawnCli");
         }
@@ -1150,8 +1384,13 @@ mod tests {
 
         let triggers = ColumnTriggersV2 {
             on_entry: Some(TriggerActionV2::SpawnCli {
-                cli: None, command: None, prompt_template: None,
-                prompt: None, flags: None, use_queue: None, model: None,
+                cli: None,
+                command: None,
+                prompt_template: None,
+                prompt: None,
+                flags: None,
+                use_queue: None,
+                model: None,
             }),
             on_exit: Some(TriggerActionV2::MoveColumn {
                 target: "next".to_string(),
@@ -1209,7 +1448,7 @@ mod tests {
 
     #[test]
     fn test_resolve_column_target_next() {
-        let (conn, ws, _, col2, col3) = setup();
+        let (conn, ws, _, col2, _) = setup();
         let task = db::insert_task(&conn, &ws.id, &col2.id, "Task", None).unwrap();
 
         let result = resolve_column_target(&conn, &task, "next").unwrap();
@@ -1219,7 +1458,7 @@ mod tests {
 
     #[test]
     fn test_resolve_column_target_previous() {
-        let (conn, ws, col1, col2, _) = setup();
+        let (conn, ws, _, col2, _) = setup();
         let task = db::insert_task(&conn, &ws.id, &col2.id, "Task", None).unwrap();
 
         let result = resolve_column_target(&conn, &task, "previous").unwrap();
@@ -1315,8 +1554,14 @@ mod tests {
         }"#;
 
         let triggers: ColumnTriggersV2 = serde_json::from_str(json).unwrap();
-        assert!(matches!(triggers.on_entry, Some(TriggerActionV2::RunScript { .. })));
-        assert!(matches!(triggers.on_exit, Some(TriggerActionV2::MoveColumn { .. })));
+        assert!(matches!(
+            triggers.on_entry,
+            Some(TriggerActionV2::RunScript { .. })
+        ));
+        assert!(matches!(
+            triggers.on_exit,
+            Some(TriggerActionV2::MoveColumn { .. })
+        ));
         assert!(triggers.exit_criteria.is_some());
         let exit = triggers.exit_criteria.unwrap();
         assert_eq!(exit.criteria_type, ExitCriteriaTypeV2::AgentComplete);
@@ -1336,7 +1581,10 @@ mod tests {
                 ..
             })
         ));
-        assert!(matches!(triggers.on_exit, Some(TriggerActionV2::SpawnCli { .. })));
+        assert!(matches!(
+            triggers.on_exit,
+            Some(TriggerActionV2::SpawnCli { .. })
+        ));
         assert_eq!(
             triggers.exit_criteria.unwrap().criteria_type,
             ExitCriteriaTypeV2::AgentComplete
@@ -1354,7 +1602,10 @@ mod tests {
         let step_type = step_json.get("type").and_then(|v| v.as_str()).unwrap();
         let step_name = step_json.get("name").and_then(|v| v.as_str()).unwrap();
         let command = step_json.get("command").and_then(|v| v.as_str()).unwrap();
-        let continue_on_error = step_json.get("continueOnError").and_then(|v| v.as_bool()).unwrap_or(false);
+        let continue_on_error = step_json
+            .get("continueOnError")
+            .and_then(|v| v.as_bool())
+            .unwrap_or(false);
 
         assert_eq!(step_type, "bash");
         assert_eq!(step_name, "Build");
@@ -1395,13 +1646,20 @@ mod tests {
     #[test]
     fn test_resolved_step_defaults() {
         // Step with minimal fields — should use defaults
-        let step_json: serde_json::Value = serde_json::from_str(
-            r#"{"type": "bash"}"#
-        ).unwrap();
+        let step_json: serde_json::Value = serde_json::from_str(r#"{"type": "bash"}"#).unwrap();
 
-        let command = step_json.get("command").and_then(|v| v.as_str()).unwrap_or("");
-        let name = step_json.get("name").and_then(|v| v.as_str()).unwrap_or("Step");
-        let continue_on_error = step_json.get("continueOnError").and_then(|v| v.as_bool()).unwrap_or(false);
+        let command = step_json
+            .get("command")
+            .and_then(|v| v.as_str())
+            .unwrap_or("");
+        let name = step_json
+            .get("name")
+            .and_then(|v| v.as_str())
+            .unwrap_or("Step");
+        let continue_on_error = step_json
+            .get("continueOnError")
+            .and_then(|v| v.as_bool())
+            .unwrap_or(false);
 
         assert_eq!(command, "");
         assert_eq!(name, "Step");
@@ -1428,7 +1686,8 @@ mod tests {
                 assert!(step_type.is_some(), "Step missing type field");
                 assert!(
                     matches!(step_type.unwrap(), "bash" | "agent" | "check"),
-                    "Invalid step type: {:?}", step_type
+                    "Invalid step type: {:?}",
+                    step_type
                 );
             }
         }
@@ -1439,9 +1698,11 @@ mod tests {
     #[test]
     fn test_exit_criteria_type_extraction() {
         // Mimics the JSON parsing in evaluate_exit_criteria
-        let triggers_json = r#"{"exit_criteria": {"type": "agent_complete", "auto_advance": true}}"#;
+        let triggers_json =
+            r#"{"exit_criteria": {"type": "agent_complete", "auto_advance": true}}"#;
         let parsed: serde_json::Value = serde_json::from_str(triggers_json).unwrap();
-        let exit_type = parsed.get("exit_criteria")
+        let exit_type = parsed
+            .get("exit_criteria")
             .and_then(|v| v.get("type"))
             .and_then(|v| v.as_str())
             .unwrap_or("manual");
@@ -1452,7 +1713,8 @@ mod tests {
     fn test_exit_criteria_missing_defaults_to_manual() {
         let triggers_json = r#"{}"#;
         let parsed: serde_json::Value = serde_json::from_str(triggers_json).unwrap();
-        let exit_type = parsed.get("exit_criteria")
+        let exit_type = parsed
+            .get("exit_criteria")
             .and_then(|v| v.get("type"))
             .and_then(|v| v.as_str())
             .unwrap_or("manual");
@@ -1463,7 +1725,8 @@ mod tests {
     fn test_max_retries_extraction() {
         let triggers_json = r#"{"exit_criteria": {"type": "agent_complete", "max_retries": 3}}"#;
         let parsed: serde_json::Value = serde_json::from_str(triggers_json).unwrap();
-        let max_retries = parsed.get("exit_criteria")
+        let max_retries = parsed
+            .get("exit_criteria")
             .and_then(|v| v.get("max_retries"))
             .and_then(|v| v.as_u64())
             .unwrap_or(0);
@@ -1472,9 +1735,11 @@ mod tests {
 
     #[test]
     fn test_auto_advance_extraction() {
-        let triggers_json = r#"{"exit_criteria": {"type": "script_success", "auto_advance": true}}"#;
+        let triggers_json =
+            r#"{"exit_criteria": {"type": "script_success", "auto_advance": true}}"#;
         let parsed: serde_json::Value = serde_json::from_str(triggers_json).unwrap();
-        let auto_advance = parsed.get("exit_criteria")
+        let auto_advance = parsed
+            .get("exit_criteria")
             .and_then(|v| v.get("auto_advance"))
             .and_then(|v| v.as_bool())
             .unwrap_or(false);
@@ -1485,7 +1750,8 @@ mod tests {
     fn test_timeout_extraction() {
         let triggers_json = r#"{"exit_criteria": {"type": "time_elapsed", "timeout": 600}}"#;
         let parsed: serde_json::Value = serde_json::from_str(triggers_json).unwrap();
-        let timeout = parsed.get("exit_criteria")
+        let timeout = parsed
+            .get("exit_criteria")
             .and_then(|v| v.get("timeout"))
             .and_then(|v| v.as_u64())
             .unwrap_or(300);
